Move page components to new pages folder
