+++ imports/ui/pages/NowPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import NowContainer from '../../../containers/NowContainer.jsx';
-import RecentRecordsContainer from '../../../containers/RecentRecordsContainer.jsx';
+import NowContainer from '../../containers/NowContainer.jsx';
+import RecentRecordsContainer from '../../containers/RecentRecordsContainer.jsx';
 
 /**
  * Component that handles the on-going record
